Added links to the html
So we published the GitHub pages site so that we can finish our E2E testing but we can't specify which folder to publish from. 

We've created a hacky solution by creating a link to the GitHub pages that we can use to navigate to the HTML source page itself.